// src/components/CardCoins/types.ts
export interface CardCoinsProps {
    askPrice: string;
    askQty: string;
    bidPrice: string;
    bidQty: string;
    closeTime: number;
    count:number;
    firstId: number;
    highPrice: string;
    lastId:number;
    lastPrice: string;
    lastQty: string;
    lowPrice: string;
    openPrice: string;
    openTime: number;
    prevClosePrice:string;
    priceChange: string;
    priceChangePercent: string;
    quoteVolume: string;
    symbol: string;
    volume: string;
    weightedAvgPrice: string;
}

export interface Props {
    data: CardCoinsProps
}